+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/bean/DefaultPipelineUrlBeanImpl.kt
@@ -27,15 +27,22 @@
 
 package com.tencent.devops.process.bean
 
-interface PipelineUrlBean {
+import com.tencent.devops.common.service.config.CommonConfig
+import com.tencent.devops.common.service.utils.HomeHostUtil
 
-    /**
-     * 生成构建详情访问链接
-     */
-    fun genBuildDetailUrl(projectCode: String, pipelineId: String, buildId: String): String
+class DefaultPipelineUrlBeanImpl constructor(private val commonConfig: CommonConfig) : PipelineUrlBean {
+    override fun genBuildDetailUrl(
+        projectCode: String,
+        pipelineId: String,
+        buildId: String,
+        needShortUrl: Boolean?
+    ): String {
+        return "${HomeHostUtil
+            .getHost(commonConfig.devopsHostGateway!!)}/console/pipeline/$projectCode/$pipelineId/detail/$buildId"
+    }
 
-    /**
-     * 生成手机侧的构建详情访问链接
-     */
-    fun genAppBuildDetailUrl(projectCode: String, pipelineId: String, buildId: String): String
+    override fun genAppBuildDetailUrl(projectCode: String, pipelineId: String, buildId: String): String {
+        return "${HomeHostUtil
+            .getHost(commonConfig.devopsHostGateway!!)}/console/pipeline/$projectCode/$pipelineId/detail/$buildId"
+    }
 }
