+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/notify/PipelineBuildNotifyListener.kt
@@ -56,7 +56,10 @@ class PipelineBuildNotifyListener @Autowired constructor(
                 event.sendReviewNotify(
                     templateCode = notifyTemplateEnumType.templateCode,
                     reviewUrl = pipelineUrlBean.genBuildDetailUrl(
-                        projectCode = event.projectId, pipelineId = event.pipelineId, buildId = event.buildId
+                        projectCode = event.projectId,
+                        pipelineId = event.pipelineId,
+                        buildId = event.buildId,
+                        needShortUrl = true
                     ),
                     reviewAppUrl = pipelineUrlBean.genAppBuildDetailUrl(
                         projectCode = event.projectId, pipelineId = event.pipelineId, buildId = event.buildId
